add vm-client stack compoenent for the vms
