+++ packages/core/engine/src/index.ts
@@ -243,6 +243,7 @@ export class Engine {
   blockchainStackComponents() {
     this.registerModulePackage('embark-blockchain', { plugins: this.plugins, ipc: this.ipc });
     this.registerModulePackage('embark-blockchain-client');
+    this.registerModulePackage('embark-vm-client');
     this.registerModulePackage('embark-process-logs-api-manager');
   }
 
@@ -256,6 +257,7 @@ export class Engine {
     this.registerModulePackage('embark-contracts-manager', { plugins: this.plugins, compileOnceOnly: options.compileOnceOnly });
     this.registerModulePackage('embark-deployment', { plugins: this.plugins, onlyCompile: options.onlyCompile });
     this.registerModulePackage('embark-blockchain-client');
+    this.registerModulePackage('embark-vm-client');
     this.registerModulePackage('embark-storage');
     this.registerModulePackage('embark-communication');
     this.registerModulePackage('embark-namesystem');
