+++ packages/stack/blockchain-client/src/index.js
@@ -9,29 +9,6 @@ class BlockchainClient {
     this.events = embark.events;
     this.config = embark.config;
 
-    this.vms = [];
-    this.events.setCommandHandler("blockchain:vm:register", (name, handler) => {
-      this.vms.push({name, handler: handler()});
-    });
-
-    this.events.setCommandHandler("blockchain:client:vmProvider", async (vmName, cb) => {
-      if (typeof vmName === 'function') {
-        cb = vmName;
-        vmName = '';
-      }
-      if (!this.vms.length) {
-        return cb(`Failed to get the VM provider. Please register one using 'blockchain:vm:register', or by ensuring the 'embark-ganache' package is registered.`);
-      }
-      if (vmName) {
-        const vm = this.getVmClient(vmName);
-        if (!vm) {
-          return cb(__("No VM of name %s registered. Please register it using using 'blockchain:vm:register'", vmName));
-        }
-        return cb(null, vm.handler);
-      }
-      return cb(null, this.vms[this.vms.length - 1].handler);
-    });
-
     this.blockchainClients = {};
     this.client = null;
     this.events.setCommandHandler("blockchain:client:register", (clientName, blockchainClient) => {
@@ -53,10 +30,6 @@ class BlockchainClient {
       }
       cb(null, provider);
     });
-  }
-
-  getVmClient(vmName) {
-    return this.vms.find(vm => vm.name === vmName);
   }
 
   async _getProvider(clientName, endpoint) {
