+++ packages/stack/vm-client/README.md
@@ -1,0 +1,6 @@
+# `embark-vm-client`
+
+> Provides ability to register a VM technology in Embark, ie Ganache.
+
+Visit [framework.embarklabs.io](https://framework.embarklabs.io/) to get started with
+[Embark](https://github.com/embarklabs/embark).
