+++ packages/stack/vm-client/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "embark-blockchain-client",
-  "version": "5.1.1-nightly.2",
+  "name": "embark-vm-client",
+  "version": "5.1.0",
   "author": "Iuri Matias <iuri.matias@gmail.com>",
   "contributors": [],
-  "description": "Provides ability to register a blockchain technology in Embark, ie Ethereum.",
-  "homepage": "https://github.com/embarklabs/embark/blob/master/packages/stack/blockchain-client#readme",
+  "description": "Provides ability to register a VM technology in Embark, ie Ganache",
+  "homepage": "https://github.com/embarklabs/embark/blob/master/packages/stack/vm-client#readme",
   "bugs": "https://github.com/embarklabs/embark/issues",
   "keywords": [
     "blockchain",
@@ -20,7 +20,7 @@
   ],
   "license": "MIT",
   "repository": {
-    "directory": "packages/stack/blockchain-client",
+    "directory": "packages/stack/vm-client",
     "type": "git",
     "url": "https://github.com/embarklabs/embark.git"
   },
@@ -46,11 +46,10 @@
   "dependencies": {
     "@babel/runtime-corejs3": "7.7.4",
     "core-js": "3.4.3",
-    "embark-i18n": "^5.1.1-nightly.0",
-    "web3": "1.2.6"
+    "embark-i18n": "^5.1.1-nightly.0"
   },
   "devDependencies": {
-    "embark-solo": "^5.1.1-nightly.2",
+    "embark-solo": "^5.1.0",
     "eslint": "5.7.0",
     "npm-run-all": "4.1.5",
     "rimraf": "3.0.0"
